+++ ChangeLog.md
@@ -3,6 +3,7 @@
 * v1.0.1:
     * Fix cascade deletion on `_etransactions_transaction` table
     * Fix non escaped description text
+    * Add free amount for product. e.g. for donation.
     
 * v1.0.0: 
     * Initial Release